requirements.txt
pywaggle[all]
pyserial